requirements.txt
altair==4.2.2
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
defusedxml==0.7.1
docxcompose==1.4.0
docxtpl==0.19.1
entrypoints==0.4
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.2
odfpy==1.4.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
soupsieve==2.6
streamlit==1.41.1
streamlit-aggrid==1.0.5
tenacity==9.0.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
